Adds hack to fix issue with webview and the use of network.
WLXWebViewReloader uses an internal WKWebView to load a Javacript
code to be able to use SocketIO JS client in order to connect
with the watcher server. The problem is that the WKWebView does
not perform network requests properly if it is not attached to
a view hierarchy.

To fix this "issue" I a attached the WLXWebViewReloader's
internal webview to the target webview's superview.